Add support for CSApprox fall-back.
This "disables" CSApprox if it's active, so it doesn't run for all
colorschemes. However, if a colorscheme matches g:csexact_blacklist or
if the terminal configuration is not supported, CSApprox is invoked to
handle the case. Currently CSApprox is also used when g:colors_name is
unset, but that may be changed soon.