Move data back to its own file.
This combination of data and utility functions was probably a bad idea in
general, but it also created a specific problem: using csexact#TermDetails would
load the autoload file, and loading the file would use g:csexact_rgbtxt to find
the rgb.txt file. This meant that if you wanted to use g:csexact_rgbtxt, you'd
have to set it before calling csexact#TermDetails. This is a silly and pointless
restriction, and may make some legitimate uses impossible (perhaps you need
terminal information to decide what rgb.txt to use--unlikely, but there's no
reason to disallow it).