Fixes for :highlight parsing.
There are some ways to create a highlight groups with an empty name (this is
likely a Vim bug). Example:

    :sign define Foo linehl=

This creates problems for the "line unwrapping" that CSExact did, where a line
beginning with space was taken to be continuations of the previous line. A group
with an empty name looks like a wrapped line.

The unwrapping was unreliable in other ways. With very narrow windows or very
long group names, the wrapping could be less predictable. It's also possible to
get the group name and the 'xxx' after it to merge into one word with no space
between, causing the normal regex matching to fail.

So instead of unwrapping, force :highlight to not wrap by temporarily setting
'columns' to something ridiculously high.

Also loosening the regex for group names a bit, because Vim is actually very
lenient in what it accepts (though it emits warnings for things that are not
letters, numbers, or underscores).