Fix highlight group name regex.
While the Vim code that validates a group name allows any printable character,
it's probably not going to include whitespace. This would probably be prevented
by the command parsing treating the whitespace as a token separator.

The old regex made the group names have trailing spaces, which surprisingly
didn't cause problems most of the time, but in a few specific cases it broke
things (e.g., looking up a group by name in a dictionary).