Handle groups with both attributes and a link.
Sometimes highlight groups have specific attributes and also link to another
group (you have to use ":highlight! link ..." to create such a link). This looks
like

    SpellLocal     xxx term=underline ctermbg=14 gui=undercurl guisp=Cyan
                   links to Error

In the past, lines that started with whitespace were combined with the previous
line, and there was a bit of special handling to allow "links to" at the end of
the line. Now we keep the lines separate for better reliability, so we have to
look ahead for a "links to" on the next line.

Before this change, colorschemes that do a "highlight! link" could produce
warnings about bad highlight lines (and it would perhaps break the group link as
well).