New GetHighlights implementation
This version no longer tries to capture and parse :highlight output,
which has caused many problems. Instead it uses hlget().

hlget() was introduced in Vim 9, so this might be raising the earliest
supported version, I'm not entirely sure where it was before.

This fixes github issue #10, changes terminal size, because it no longer
sets 'columns'.